requirements.txt
streamlit==1.51.0
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2 
matplotlib==3.7.2
seaborn==0.12.2
Pillow==12.0.0
joblib==1.3.2  # Para futuros modelos